+++ src/main/java/net/minestom/server/event/EventNodeImpl.java
@@ -17,6 +17,8 @@ import java.util.function.BiPredicate;
 import java.util.function.Consumer;
 
 non-sealed class EventNodeImpl<T extends Event> implements EventNode<T> {
+    private static final boolean ALLOW_MULTIPLE_PARENTS = Boolean.getBoolean("minestom.event.multiple-parents");
+
     static final Object GLOBAL_CHILD_LOCK = new Object();
 
     private final Map<Class, Handle<T>> handleMap = new ConcurrentHashMap<>();
@@ -106,7 +108,7 @@ non-sealed class EventNodeImpl<T extends Event> implements EventNode<T> {
     public @NotNull EventNode<T> addChild(@NotNull EventNode<? extends T> child) {
         synchronized (GLOBAL_CHILD_LOCK) {
             final var childImpl = (EventNodeImpl<? extends T>) child;
-            Check.stateCondition(childImpl.parent != null, "Node already has a parent");
+            Check.stateCondition(!ALLOW_MULTIPLE_PARENTS && childImpl.parent != null, "Node already has a parent");
             Check.stateCondition(Objects.equals(parent, child), "Cannot have a child as parent");
             if (!children.add((EventNodeImpl<T>) childImpl)) return this; // Couldn't add the child (already present?)
             childImpl.parent = this;
@@ -217,6 +219,7 @@ non-sealed class EventNodeImpl<T extends Event> implements EventNode<T> {
 
     @Override
     public @Nullable EventNode<? super T> getParent() {
+        Check.stateCondition(ALLOW_MULTIPLE_PARENTS, "Cannot use getParent when multiple parents are allowed");
         return parent;
     }