+++ CHANGELOG.md
@@ -10,3 +10,4 @@ Some of these are pending, some deserve PRs, others are just minor tweaks
 * **breaking** Replace permission system with a simple user pluggable alternative
 * **breaking** Remove tinylog and MinestomTerminal implementation
 * Add `Tag.Transient`
+* Optionally allow multiple parents in event nodes